% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_laplace_NR.R
\name{calculate_posterior_VL}
\alias{calculate_posterior_VL}
\title{Vecchia Laplace extendion of GPVecchia}
\usage{
calculate_posterior_VL(z, vecchia.approx,
  likelihood_model = c("gaussian", "logistic", "poisson", "gamma",
  "beta", "gamma_alt"), covparms, likparms = list(alpha = 2, sigma =
  sqrt(0.1)), max.iter = 50, convg = 1e-06, return_all = FALSE,
  y_init = NA, prior_mean = rep(0, length(z)))
}
\arguments{
\item{vecchia.approx:}{a vecchia object as generated by vecchia_specify()}

\item{likelihood_model:}{text describing likelihood model to be used for observations}

\item{covparms:}{covariance parameters as a vector}

\item{likparms:}{likelihood parameters for the likelihood_model, as a list}

\item{covmodel:}{covariance model, 'matern' by default}
}
\value{
(multivariate normal) loglikelihood implied by the Vecchia approximation
}
\description{
Vecchia Laplace extendion of GPVecchia
}
\examples{
z=rnorm(5); locs=matrix(1:5,ncol=1); vecchia.approx=vecchia_specify(locs,m=5)
vecchia_likelihood(z,vecchia.approx,covparms=c(1,2,.5),nuggets=.2)
}
