% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRA_utility-functions.r
\name{getMatCov}
\alias{getMatCov}
\title{extract the required elements from the covariance matrix}
\usage{
getMatCov(V, Sigma)
}
\arguments{
\item{V}{the object returned by vecchia_specify}

\item{Sigma}{The full covariance matrix}
}
\value{
matrix of size n x (m+1) with only those elements that 
are used by the incomplete Cholesky decomposition
}
\description{
This function takes the entire covariance matrix and creates
a matrix of covariances based on the vecchia approximatino object
}
