% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMatCovFromFactorCpp}
\alias{getMatCovFromFactorCpp}
\title{Calculate the covariance values required by HV for
matrix factors passed as sparse matrices}
\usage{
getMatCovFromFactorCpp(F, revNNarray)
}
\arguments{
\item{F}{factor of a matrix in a sparse format}

\item{revNNarray}{array with the neighbourhood structure}
}
\value{
matrix with covariance values
}
\description{
Calculate the covariance values required by HV for
matrix factors passed as sparse matrices
}
