\name{GPvecchia-package}
\alias{GPvecchia-package}
\alias{GPvecchia}
\docType{package}
\title{
\packageTitle{GPvecchia}
}
\description{
\packageDescription{GPvecchia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GPvecchia}
\packageIndices{GPvecchia}
temp
}
\author{
\packageAuthor{GPvecchia}

Maintainer: \packageMaintainer{GPvecchia}
}
\references{
temp
}

\keyword{ package }
\seealso{
temp
}
\examples{
temp
}
