% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPvecchia.R
\docType{package}
\name{GPvecchia}
\alias{GPvecchia}
\alias{GPvecchia-package}
\title{GPvecchia: fast, scalable Gaussian process approximations}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
