% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaternFun}
\alias{MaternFun}
\title{Calculate Matern covariance function}
\usage{
MaternFun(distmat, covparms)
}
\arguments{
\item{distmat}{A matrix with distances between points}

\item{covparams}{A vector with parameters (marg. variance, range, smoothness)}
}
\description{
Calculate Matern covariance function
}
