% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_prediction.R
\name{SelInv}
\alias{SelInv}
\title{selected inverse of a sparse matrix}
\usage{
SelInv(cholmat)
}
\arguments{
\item{cholmat}{cholesky factor L of a positive definite sparseMatrix A}
}
\value{
sparse inverse of A, with same sparsity pattern as L
}
\description{
selected inverse of a sparse matrix
}
\examples{
A=Matrix::sparseMatrix(1:9,1:9,x=4); L=chol(A)
SelInv(L)
}
