% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_prediction.R
\name{V2covmat}
\alias{V2covmat}
\title{compute covariance matrix from V.ord}
\usage{
V2covmat(preds, vecchia.approx)
}
\arguments{
\item{vecchia.approx}{Results from vecchia_specify}

\item{preds:}{Results from vecchia_prediction @MK}
}
\value{
Covariance matrix of prediction. Do not do this for large n or n.p!!!
}
\description{
compute covariance matrix from V.ord
}
\examples{
V2covmat=function(preds,vecchia.approx)
}
