% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_prediction.R
\name{vecchia_lincomb}
\alias{vecchia_lincomb}
\title{linear combination of predictions
compute the distribution of a linear combination Hy}
\usage{
vecchia_lincomb(H, vecchia.approx, V.ord, cov.mat = FALSE)
}
\arguments{
\item{H:}{sparse matrix with n.all columns specifying the linear combination}

\item{vecchia.approx:}{a vecchia object as generated by vecchia_specify()}

\item{V.ord:}{ordered V matrix from vecchia_prediction() or U2V()}

\item{cov.mat:}{logical TRUE or FALSE -- should the entire covariance matrix be returned (only do if H has a small number of rows)}
}
\value{
Variance of linear combination of predictions.
}
\description{
linear combination of predictions
compute the distribution of a linear combination Hy
}
\examples{
z=rnorm(5); locs=matrix(1:5,ncol=1); vecchia_specify=function(z,locs,m=5,locs.pred=(1:5)+.5)
preds=vecchia_prediction=function(vecchia.approx,covparms=c(1,2,.5),nuggets=.2)
H=sparseMatrix(i=rep(1,n.p),j=n+(1:n.p),x=1/n.p)
vecchia_lincomb(H,vecchia.approx,preds$V.ord,cov.mat=TRUE)
}
