% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_prediction.R
\name{vecchia_prediction}
\alias{vecchia_prediction}
\title{Vecchia prediction}
\usage{
vecchia_prediction(vecchia.approx, covparms, nuggets, var.exact,
  covmodel = "matern")
}
\arguments{
\item{vecchia.approx}{@ MK}

\item{var.exact}{@ MK}

\item{covparms:}{covariance parameters as a vector}

\item{nuggets:}{nugget}

\item{covmodel:}{covariance model, 'matern' by default.}
}
\value{
Posterior mean, obverved mean, posterior variance, observed variance and V matrix.
}
\description{
Vecchia prediction
}
\examples{
vecchia_prediction=function(vecchia.approx,covparms,nuggets,var.exact,covmodel='matern')
}
