% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_specify.R
\name{vecchia_specify}
\alias{vecchia_specify}
\title{specify the vecchia approximation}
\usage{
vecchia_specify(z, locs, m, ordering, cond.yz, locs.pred, ordering.pred,
  pred.cond, conditioning)
}
\arguments{
\item{locs:}{nxd matrix of obs locs}

\item{ordering:}{options are 'coord' or 'maxmin'}

\item{cond.yz:}{options are 'y', 'z', 'SGV', 'SGVT', and 'zy'}

\item{ordering.pred:}{options are 'obspred' or 'general'}

\item{pred.cond:}{prediction conditioning, options are 'general' or 'independent'}

\item{conditioning:}{conditioning on 'NN' (nearest neighbor) or 'firstm' (fixed set for low rank)}
}
\value{
An object that specifies the vecchia approximation.@MK
}
\description{
specify the vecchia approximation, prepare U. This fct does not depend on parameter values, only has to be run once before repeated likelihood evals
}
\examples{
vecchia_specify=function(z,locs,m,ordering,cond.yz,locs.pred,ordering.pred,pred.cond,conditioning)
}
