% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createU.R
\name{createL}
\alias{createL}
\title{create the sparse triangular L matrix for specific parameters}
\usage{
createL(vecchia.approx, covparms, covmodel)
}
\arguments{
\item{vecchia.approx}{object returned by \code{\link{vecchia_specify}}}

\item{covparms}{vector of covariance parameters}

\item{covmodel}{covariance model. currently implemented:}

\item{nuggets}{nugget variances -- if a scalar is provided, variance is assumed constant}
}
\value{
list containing the sparse lower triangular L,
}
\description{
create the sparse triangular L matrix for specific parameters
}
\examples{
z=rnorm(9); locs=matrix(1:9,ncol=1); vecchia.approx=vecchia_specify(locs,m=5)
L = createL(vecchia.approx, covparms=c(1,2,.5), 'matern')
}
