\name{GPvecchia-package}
\alias{GPvecchia-package}
\alias{GPvecchia}
\docType{package}
\title{
\packageTitle{GPvecchia}
}
\description{
\packageDescription{GPvecchia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GPvecchia}
\packageIndices{GPvecchia}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GPvecchia}

Maintainer: \packageMaintainer{GPvecchia}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
