% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPvecchia.R
\docType{package}
\name{GPvecchia}
\alias{GPvecchia}
\alias{GPvecchia-package}
\title{GPvecchia: fast, scalable Gaussian process approximations}
\description{
The package can be used for parameter inference and prediction for Gaussian and non-Gaussian spatial data using many popular GP approximation methods.
}
