% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_prediction.R
\name{V2covmat}
\alias{V2covmat}
\title{compute covariance matrix from V.ord
Do not run this function for large n or n.p!!!}
\usage{
V2covmat(preds)

V2covmat(preds)
}
\arguments{
\item{preds:}{Object returned by vecchia_prediction()}

\item{preds:}{Object returned by vecchia_prediction()}
}
\value{
Covariance matrix at all locations in original order

Covariance matrix at all locations in original order
}
\description{
compute covariance matrix from V.ord
Do not run this function for large n or n.p!!!

compute covariance matrix from V.ord
Do not run this function for large n or n.p!!!
}
\examples{
z=rnorm(5); locs=matrix(1:5,ncol=1); vecchia_specify=function(z,locs,m=5,locs.pred=(1:5)+.5)
preds=vecchia_prediction(vecchia.approx,covparms=c(1,2,.5),nuggets=.2)
V2covmat=function(preds)
z=rnorm(5); locs=matrix(1:5,ncol=1); vecchia_specify=function(z,locs,m=5,locs.pred=(1:5)+.5)
preds=vecchia_prediction(vecchia.approx,covparms=c(1,2,.5),nuggets=.2)
V2covmat=function(preds)
}
