% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_likelihood.R
\name{vecchia_likelihood}
\alias{vecchia_likelihood}
\title{evaluation of the likelihood}
\usage{
vecchia_likelihood(vecchia.approx, covparms, nuggets, covmodel = "matern")
}
\arguments{
\item{vecchia.approx}{Results from vecchia_prediction @MK}

\item{covparms:}{covariance parameters as a vector}

\item{nuggets:}{nugget}

\item{var.exact}{@MK}

\item{covmodel:}{covariance model, 'matern' by default.}
}
\value{
An object that specifies the vecchia approximation. @MK
}
\description{
evaluation of the likelihood
}
\examples{
vecchia_likelihood=function(vecchia.approx,covparms,nuggets,covmodel='matern')
}
