% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecchia_prediction.R
\name{vecchia_lincomb}
\alias{vecchia_lincomb}
\title{linear combination of predictions}
\usage{
vecchia_lincomb(H, vecchia.approx, V.ord, cov.mat = FALSE)
}
\arguments{
\item{vecchia.approx}{@ MK}

\item{V.ord}{ordered V matrix}

\item{cov.mat}{logical TRUE or FALSE}

\item{H:}{@MK}
}
\value{
Variance of linear combination of predictions.
}
\description{
linear combination of predictions
}
\examples{
vecchia_lincomb=function(H,vecchia.approx,V.ord,cov.mat=FALSE)
}
